# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flipzy API",
        "description": "Flipzy Property Management Platform API - Complete API documentation for property listings, messaging, analytics, and admin features.\n\n## Authentication\n\nThis API uses Laravel Passport for authentication. To authenticate:\n\n1. Use the `/register` or `/login` endpoint to get an access token\n2. Copy the `access_token` from the response\n3. Click the 'Authorize' button at the top of this page\n4. Enter: `Bearer YOUR_ACCESS_TOKEN` (replace YOUR_ACCESS_TOKEN with your actual token)\n5. Click 'Authorize' and then 'Close'\n\nAll protected endpoints will now use this token automatically.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/admin/transactions": {
            "get": {
                "tags": [
                    "Admin - Transactions"
                ],
                "summary": "List all transactions (Admin only)",
                "description": "Get a paginated list of all transactions with optional filtering and sorting",
                "operationId": "9f18979450c42801220a4a597a407ee5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "completed",
                                "failed",
                                "refunded",
                                "partially_refunded"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "payment",
                                "refund",
                                "subscription",
                                "one_time"
                            ]
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/transactions/stats": {
            "get": {
                "tags": [
                    "Admin - Transactions"
                ],
                "summary": "Get transaction statistics (Admin only)",
                "description": "Get aggregated statistics about transactions including totals, counts, and trends",
                "operationId": "434ad9cc0de3318b5d523c19d5bbfaf6",
                "parameters": [
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/transactions/{transaction}": {
            "get": {
                "tags": [
                    "Admin - Transactions"
                ],
                "summary": "Get transaction details (Admin only)",
                "description": "Get detailed information about a specific transaction",
                "operationId": "ccdc2d8d447f105d496cce1788459b5d",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "Transaction UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction retrieved successfully"
                    },
                    "404": {
                        "description": "Transaction not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your access token in the format: Bearer YOUR_ACCESS_TOKEN\n\nGet your token by:\n- Registering a new account via POST /register\n- Logging in via POST /login\n\nThe token will be returned in the response as 'access_token'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}